Added a remove video player button

+++ src/helpers/data.rs
@@ -1,7 +1,7 @@
 use gtk::prelude::ObjectExt;
-use std::sync::atomic::{AtomicU64, Ordering};
+use std::sync::atomic::{AtomicU32, Ordering};
 
-static NEXT_ID: AtomicU64 = AtomicU64::new(1);
+static NEXT_ID: AtomicU32 = AtomicU32::new(1);
 
 pub fn store_data<T: 'static>(widget: &impl ObjectExt, key: &str, value: T) {
     unsafe {
@@ -13,7 +13,7 @@ pub fn get_data<T: 'static>(widget: &impl ObjectExt, key: &str) -> Option<std::p
     unsafe { widget.data::<T>(key) }
 }
 
-pub fn get_next_id() -> u64 {
+pub fn get_next_id() -> u32 {
     NEXT_ID.fetch_add(1, Ordering::Relaxed)
 }
